7-1

+++ library/stack.h
@@ -12,7 +12,7 @@ typedef enum{
     ERROR =0
 }Stack_err_t;
 
-typedef double ElemType; 
+typedef int ElemType; 
 #define STACK_INIT_SIZE 100
 #define STACK_INCREMENT 10
